Renamed database file. Rearranged use case package slightly and made semantical enchancements.
