+++ app/src/main/java/ftn/com/trippleaproject/repository/remote/RemoteRepositoryModule.java
@@ -4,7 +4,7 @@ import javax.inject.Singleton;
 
 import dagger.Module;
 import dagger.Provides;
-import ftn.com.trippleaproject.usecase.crud.repository.remote.NewsArticleRemoteDao;
+import ftn.com.trippleaproject.usecase.repository.dependencies.remote.NewsArticleRemoteDao;
 import ftn.com.trippleaproject.repository.remote.dao.mock.NewsArticleRemoteDaoMock;
 
 @Module
