+++ app/src/main/java/ftn/com/trippleaproject/usecase/UseCaseModule.java
@@ -4,17 +4,17 @@ import javax.inject.Singleton;
 
 import dagger.Module;
 import dagger.Provides;
-import ftn.com.trippleaproject.usecase.crud.repository.remote.NewsArticleRemoteDao;
-import ftn.com.trippleaproject.usecase.crud.NewsArticleCrudUseCase;
-import ftn.com.trippleaproject.usecase.crud.repository.local.NewsArticleLocalDao;
+import ftn.com.trippleaproject.usecase.repository.dependencies.remote.NewsArticleRemoteDao;
+import ftn.com.trippleaproject.usecase.repository.NewsArticleUseCase;
+import ftn.com.trippleaproject.usecase.repository.dependencies.local.NewsArticleLocalDao;
 
 @Module
 public class UseCaseModule {
 
     @Provides
     @Singleton
-    NewsArticleCrudUseCase provideNewsArticleCrudUseCase(NewsArticleRemoteDao newsArticleRemoteDao,
-                                                         NewsArticleLocalDao newsArticleLocalDao) {
-        return new NewsArticleCrudUseCase(newsArticleRemoteDao, newsArticleLocalDao);
+    NewsArticleUseCase provideNewsArticleCrudUseCase(NewsArticleRemoteDao newsArticleRemoteDao,
+                                                     NewsArticleLocalDao newsArticleLocalDao) {
+        return new NewsArticleUseCase(newsArticleRemoteDao, newsArticleLocalDao);
     }
 }
