+++ app/src/main/java/ftn/com/trippleaproject/usecase/repository/NewsArticleUseCase.java
@@ -1,22 +1,22 @@
-package ftn.com.trippleaproject.usecase.crud;
+package ftn.com.trippleaproject.usecase.repository;
 
 import org.reactivestreams.Subscriber;
 
 import java.util.List;
 
 import ftn.com.trippleaproject.domain.NewsArticle;
-import ftn.com.trippleaproject.usecase.crud.repository.remote.NewsArticleRemoteDao;
-import ftn.com.trippleaproject.usecase.crud.repository.local.NewsArticleLocalDao;
+import ftn.com.trippleaproject.usecase.repository.dependencies.remote.NewsArticleRemoteDao;
+import ftn.com.trippleaproject.usecase.repository.dependencies.local.NewsArticleLocalDao;
 import io.reactivex.Flowable;
 import io.reactivex.schedulers.Schedulers;
 
-public class NewsArticleCrudUseCase {
+public class NewsArticleUseCase {
 
     private final NewsArticleRemoteDao newsArticleRemoteDao;
 
     private final NewsArticleLocalDao newsArticleLocalDao;
 
-    public NewsArticleCrudUseCase(NewsArticleRemoteDao newsArticleRemoteDao, NewsArticleLocalDao newsArticleLocalDao) {
+    public NewsArticleUseCase(NewsArticleRemoteDao newsArticleRemoteDao, NewsArticleLocalDao newsArticleLocalDao) {
         this.newsArticleRemoteDao = newsArticleRemoteDao;
         this.newsArticleLocalDao = newsArticleLocalDao;
     }
